pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "contentstack-competitive-intel"
version = "0.1.0"
description = "Agentic competitive intelligence system for Contentstack product marketing"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
